+++ servers/ktor/src/main/kotlin/com/quietmetrix/server/routes/Routing.kt
@@ -4,10 +4,12 @@ import com.quietmetrix.server.config.AppConfig
 import io.ktor.server.routing.Routing
 
 fun Routing.configureAllRoutes(config: AppConfig) {
+    configureMetaRoutes()
     configureTrackRoutes()
     configureAuthRoutes(config)
     configureUserRoutes(config)
     configureProjectRoutes()
+    configureTokenRoutes()
     configureFunnelRoutes()
     configureDashboardRoutes()
     configureStaticDashboardRoutes()